Changed HAL to use generated files
This allows a script to create bindings from FIRST's HAL sources. In
HAL-Base, the generated files are located in the Generated folder,
suffixed with Generated. If any methods are needed to extend the HAL,
please place them in their respective class in the root of the HAL-Base
project.  The classes are marked partial, so they can be split between 2
file.

The script also generates the bindings that go in HAL-RoboRIO. This
means the entire interop portion of the HAL is completely scripted, and
should not need to be updated unless the allwpilib HAL is updated, at
which point we can run the script again.

The script is located in robotdotnet-tools under HAL-Generator. Its
specific to my computer, so if you want to run it you will have to
modify it.