Updated robot Main name
Note that this is a breaking commit. Any old test programs will now no
longer work. This is because if you were to create a static object in
your robot program , it would run before robot prestart would. this
would cause the HAL and the reporters to not be initialized. The only 2
fixes for this were to either force teams to initialize static objects
in the constructor, or move the Main function to a seperate class.
Moving it to a seperate class seemed like the better choice, and more in
line with the normal DotNet way.