More documentation updates
Using InheritDoc much more. This means that compiling from source does
not included some of the documentation in intellisense. In order for it
to be loaded in intellisense, you must use sandcastle to generate the
docs, which will generate an xml with the correct documentation. This
can then be copied to the DLL directory, and will show up in
intellisense. The release builder automatically does this for us, so
Nuget will have the correct documentation.