Updates to HAL and RobotBase
There are 2 big changes in here. 1st, the program now determines if it
is on the RoboRIO using a file, instead of just guessing if we are on
linux. This will allow the simulator to be ran successfully on Linux.
The second change is to allow the user to send a type to start, and not
just an assembly. If sending an assembly, and a team has 2 classes that
extend from RobotBase, it will be indeterminite which one it starts. So
instead, you can pass robotType to be your robot type, and it will start
that.