Made all HAL Rio and Sim functions internal V2
Added an InternalsVisableTo attribute to both projects. This makes it so
HAL-Base and WPILib.Test can see the functions, but WPILib and user
programs cannot. Also it means documentation will not be generated,
which is helpful.

Since we grab the methods for which Hal Impl we are using, people that
need to access the Hal should always go through HAL-Base. You should
never need to access the individual functions in the implementation.

HAL-Simulation still has some public functions, but they are for
interfacing with the simulator, and not just the HAL. So they need to
stay public. I'm actually thinking about moving these to another
assembly, but I think that might be going a little too far.