Forces DS to always show teleop enabled
Starting next season, we will be able to use the official DS. But for
now, the DS implementation we had was REALLLYYY bad, and would cause all
kind of dictionary errors, which was actually the reason for the errors
in the first place. For now, we will just force teleop enabled unless
otherwise needed.