Makes AttributedRobot its own base class
Replaces IterativeRobot with RobotBase, and creates its own control
loop. This enables us to make the public functions for teams to use the
same names as all the other robot bases. I don't know if I like using an
_ prefix for the shadow methods, but it was the best I could come up
with for now, and its easy to change later. The rest of the WPILib uses
underscores for shadow methods so it matches.