Fixes documentation build errors
When using the nuget sandcastle packages, the documentation wouldn't
build remarks or code examples. Also, if you did not have the full
sandcastle installed, VS wouldnt see it and wouldn't build it anyway, so
there was not much of a real reason to have it in the solution. This
removed it from the solution, added a batch script in the root that will
build the documentation and grab the nuget packages if people want to
build the xml docs. However normal builds will not anymore. You can also
build the AppVeyor build to build them.