Switches HAL reference loading order
Ran into an issue in the past where running the code on the roborio from
a different directory would load an unexpected dll. This forces the code
to load from the directory HAL-Base is located in, and then for load if
that fail. The directory loading will not work if the directory has
spaces in it, but that should not happen.