Removes auto downloader for the HAL
Was too much of a pain. Required too many steps to update the HAL file.
More then just downloading it and storing it in the repository. Plus I
needed a custom build until they fix the image upstream, and there was
no way in the old way to do that.